Refactor completo y nuevas funcionalidades implementadas en RoutineFragment:
- Sustituida la función touch en todas las views por la nueva accesible globalmente.
- Reemplazado RecyclerView para usar DiffUtil y optimizar actualizaciones.
- Adaptado el manejo de datos a arquitectura MVVM, separando UI y lógica.
- Añadido control de estados para la UI.
- Incorporado título de la rutina en la pantalla.
- Añadido TextView "Atrás" al lado del botón para retroceder.
- Implementada función para retroceder al fragmento anterior.
- Añadida función para mostrar PopupMenu con opciones "Ver información" y "Eliminar" al clicar el icono de tres botones.
- Añadida función para eliminar ejercicios.
- Integrado Flow para recargar cambios automáticamente.
- Añadido listado para reordenar ejercicios al pulsar EDITAR.
- Implementada lógica completa para reordenar ejercicios.
- Añadida función "Ver información" detallada para ejercicios.
- Mejoradas y extendidas descripciones de ejercicios.
- Añadida función para buscar ejercicios en YouTube al clicar el logo de YouTube.

+++ app/src/main/java/com/cursointermedio/myapplication/ui/week/WeekFragment.kt
@@ -362,7 +362,7 @@ class WeekFragment @Inject constructor() : Fragment() {
     private fun navigateToRoutine(routineId: Long) {
         findNavController().navigate(
             WeekFragmentDirections.actionWeekFragmentToRoutineFragment(
-                routineId.toInt()
+                routineId
             )
         )
     }
